Check only the interfaces in each bondgroup for liveness
If we check all interfaces for liveness, we end up bringing up all
live interfaces ever mentioned in a bonding whether or not they are
present in the bondgroup we are checking.

So check only the interfaces in the bondgroup under consideration for
liveness. (This is technically less efficient but no overhead is
incurred unless we are already using a fallback bondgroup, and even then
it is distinctly minor unless you have dozens of them, almost all down.)